Bon, j'ai tout changer...

+++ Events/Client_Messages/Message_Create.ts
@@ -1,4 +1,4 @@
-import {prefix} from '../../Utils/Misc/Settings.json';
+import {prefix, auto_delete_message} from '../../Utils/Misc/Settings.json';
 import type {ClientAttributes} from '../../Types/Client'
 import {Message} from 'discord.js-selfbot-v13';
 import 'dotenv/config';
@@ -21,7 +21,8 @@ export default {
     if (cmd) {
       if (cmd.args == true && !args.length)
           return message.reply(`**Vous devez spécifier des arguments pour votre commande !**\nPour plus de détail, executez \`${prefix}help ${cmdName}\``)
-      message.delete()
+      if (auto_delete_message == true)
+        message.delete()
       cmd.run(client, message, args)
     }
     else return message.reply(`**Cette commande n'existe pas !**\nPour obtenir la liste des commandes, veuillez executer la commande \`${prefix}help\``)
